Update License to GPL v3
I am not sure exactly how necessary this is. I am not a legal expert and can't make the call of whether this code is "mere aggregation" or "combining two or more modules into one", but was asked to change it just in case, so whatever. The license doesn't really matter to me since I just want to play a game that can be modded more easily.