➕ Add prettier/@typescript-eslint to ensure no eslint conflicts
➕ Add .editorconfig and fix .md format

➕ Add jest (no tests yet)

♻️ Prettier tsconfig.json

♻️ TS: consistent casing to avoid platform issues

♻️ tsconfig split to build and testable, build aliased to tsconfig.json

Just for simplicity, tsconfig.json just pulls in the build
config. Avoids having to mess with webpack any more than we
need to - it's supplied by Vortex, so we don't wanna touch it.

I'm sure it'd be possible to have webpack ignore the tests (or
maybe it even does it already) but it's not our build tool, it's
Vortex's.

➕ Add @jest/types

♻️ completely bananas stuff