✨♻️ Unwrap extra toplevel dir + lots of refactors
The `wrapTestSupported` and `wrapInstall` functions now do unwrapping
on giftwrapped mods. They unwrap before calling the wrapped function
and then fix the instructions to 'rewrap' the source paths.

In the case of the `internalPipelineInstaller`, this only happens for
the pipeline installer *itself*. The other installers in the pipeline
it manages are called without duplicate wrapping.