➖ Replace Prettier with Eslint rules
Got tired of the couple things prettier gets wrong in a way
that makes reading and diffs harder.

Replicated the prettier formatting style in Eslint as close
as possible.

Should be easier to layout arrays etc. now, but some things
may still be missing or not quite what they should be, so,
feel free to adjust the rules.