♻️🖼️ `ModInfo.installingDir` implemented as Path type
The `Path` type encapsulates the relative and on-disk paths for
a given path. This both allows passing the data along where it's
needed, as well as adds type safety to ensure we're using the right
kind of path in the right place.

Some work remains to properly use it, tracked in https://github.com/E1337Kat/cyberpunk2077_ext_redux/issues/251