♻️  Separate static & delayed dynamic features for main()
Vortex doesn't allow us to access the api during extension registration,
so the features are split into the constituent static and dynamic parts,
of which the dynamic are added so that they don't need to be touched until
the api is available.

The naming is really -.-, especially `vortexExt` and `vortexLib`, but we'll
have to return to it later.