🐛 Stop LO if LO File Exists but Doesn't Decode
If the file isn't there, could be legit (or first time) and we can overwrite.

But if file is there and can't be decoded, overwriting it would lose the
old order and the problem could be because of a bug.. so we pop a message.

Additional work needs to go to improve the write side to avoid error conditions.