Recursive search for AMM files
As stated in the `canon` blurb, both `.lua` and `.json` files in AMM content packs can be in any number of appropriate subdirectories. The AMM content matcher should not use the `findDirect` function for these, as it will fail to find content if nested one subdirectory away. Instead, the `findAll` function gets the job done.